Don't block IIS whilst sending request to redis.
We use concurrency::task to do it on the thread pool instead. This does
mean we eat up a thread whilst all 4 redis operations are completed. This
is _probably_ fine, but we could look at doing something smarter when the
hiredis iocp stuff is merged in. (The iocp stuff is currently on a branch
as it makes things too complicated).